requirements.txt
asgiref @ file:///work/ci_py311/asgiref_1676843199969/work
beautifulsoup4 @ file:///home/conda/feedstock_root/build_artifacts/beautifulsoup4_1680888073205/work
Django @ file:///work/ci_py311/django_1676851927387/work
django-bootstrap5 @ file:///home/conda/feedstock_root/build_artifacts/django-bootstrap5_1686162285918/work
django-braces==1.15.0
Pillow @ file:///croot/pillow_1696580024257/work
python-decouple==3.8
soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1693929250441/work
sqlparse @ file:///croot/sqlparse_1690901927333/work
